Latest stable version.
Lots of changes. Variables now marked with _ or __ at the start are varying levels of visibility in data outputs/processing. The input selection and table information is significantly improved and stores the raw data, not just the text in the table. Report1 and 2 have been bug fixed and improved. Some improvements in the layout/gui information. General bug fixes.